Factor out a dedicated pytest fixtures module.
